app/requirements.txt
gunicorn==20.0.4
Flask==2.1.
Flask-Cors==3.0.8
sqlalchemy==1.3.13
pymysql==0.9.3
google-cloud-storage==1.26.0
google-cloud-core>=1.4.1
google-cloud-secret-manager==0.1.1
requests==2.22.0
argon2-cffi==19.2.0
google-cloud-tasks==2.0.0
google-resumable-media==0.5.1
google-cloud-bigquery==1.27.2
pyarrow==9.0.0
pandas==1.3.5
gevent==21.12.0
flask_restful
pytz==2018.7
isort==5.10.1
jsonschema==4.9.1
result==0.8.0
uuid==1.30
#  Flask-2.1.0 Flask-Cors-3.0.8 Jinja2-3.1.2 MarkupSafe-2.1.1 Six-1.16.0 Werkzeug-2.2.2 aniso8601-9.0.1 argon2-cffi-19.2.0 attrs-22.1.0 cachetools-4.2.4 certifi-2022.6.15 cffi-1.15.1 chardet-3.0.4 click-8.1.3 flask-restful-0.3.9 gevent-21.12.0 google-api-core-1.32.0 google-auth-1.35.0 google-cloud-bigquery-1.27.2 google-cloud-core-1.7.3 google-cloud-secret-manager-0.1.1 google-cloud-storage-1.26.0 google-cloud-tasks-2.0.0 google-resumable-media-0.5.1 googleapis-common-protos-1.56.4 greenlet-1.1.2 grpc-google-iam-v1-0.12.4 grpcio-1.47.0 gunicorn-20.0.4 idna-2.8 importlib-metadata-4.12.0 importlib-resources-5.9.0 isort-5.10.1 itsdangerous-2.1.2 jsonschema-4.9.1 libcst-0.4.7 mypy-extensions-0.4.3 numpy-1.21.6 packaging-21.3 pandas-1.3.5 pkgutil-resolve-name-1.3.10 proto-plus-1.22.0 
# protobuf-3.20.1 pyarrow-9.0.0 pyasn1-0.4.8 pyasn1-modules-0.2.8 pycparser-2.21 pymysql-0.9.3 pyparsing-3.0.9 pyrsistent-0.18.1 python-dateutil-2.8.2 pytz-2018.7 pyyaml-6.0 requests-2.22.0 result-0.8.0 rsa-4.9 sqlalchemy-1.3.13 typing-extensions-4.3.0 typing-inspect-0.7.1 urllib3-1.25.11 uuid-1.30 zipp-3.8.1 zope.event-4.5.0 zope.interface-5.4.0